% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.fmf.R
\name{read.fmf}
\alias{read.fmf}
\alias{read.fmf.frame}
\alias{read.fmf.header}
\title{Read FlyMovieFormat (fmf) files}
\usage{
read.fmf.frame(x, i, h = NULL, what = c("raw", "integer"))

read.fmf.header(x)
}
\arguments{
\item{x}{Path to fmf file (or for \code{read.fmf.frame} an open connection).}

\item{i}{Integer (1-indexed) indicating the frame to read}

\item{h}{Optional header information in the form returned by
\code{read.fmf.header}}

\item{what}{Character vector indicating whether to return the frame as a raw
or integer matrix}
}
\description{
\code{read.fmf.frame} reads a single frame from an fmf file
  using the header information to seek to the appropriate file location.

\code{read.fmf.header} reads the initial header information of
  an fmf file including the number of frames and the frame size.
}
\references{
see \url{http://code.astraw.com/projects/motmot/fmf-format-spec.html}
}
