% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.fmf.R
\name{read.fmf.timestamps}
\alias{read.fmf.timestamps}
\title{Read the timestamp information from all frames in an fmf file}
\usage{
read.fmf.timestamps(x)
}
\arguments{
\item{x}{Path to fmf file}
}
\value{
Numeric vector of timestamps (seconds in current epoch)
}
\description{
Read the timestamp information from all frames in an fmf file
}

