% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ufmf.R
\name{read.ufmf.header}
\alias{read.ufmf.header}
\title{Read header for a ufmf (compressed FlyMovieFormat) movie}
\usage{
read.ufmf.header(x)
}
\arguments{
\item{x}{Path to ufmf file or a connection (which will not be closed)}
}
\value{
a list containing fields describing the location of frames in the
movie file.
}
\description{
Read header for a ufmf (compressed FlyMovieFormat) movie
}

