% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmfio-package.R
\docType{package}
\name{fmfio-package}
\alias{fmfio-package}
\alias{fmfio}
\title{Read Fly Movie Format image time series}
\description{
Read Fly Movie Format image time series
}
\details{
There are two FMF formats: \itemize{

\item{\bold{fmf}}{ A very simple uncompressed movie format.}

\item{\bold{ufmf}}{ Compressed format based on background subtraction that can
provide very significant space savings.}

}
}
\seealso{
\code{\link{read.fmf}}, \code{\link{read.ufmf}}
}
\keyword{package}
