% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ufmf.frame.R
\name{read.ufmf}
\alias{read.ufmf}
\title{Read frame from a ufmf format movie}
\usage{
read.ufmf(x, framei = NULL)
}
\arguments{
\item{x}{Either a path to a file on disk, a \code{\link{connection}} or a
parsed ufmf header object (as returned by \code{\link{read.ufmf.header}})}

\item{framei}{Integer index of the frame to read}
}
\value{
A list containing the following elements \itemize{

  \item{im} The requested video frame

  \item{header} The header, which may have been modified

  \item{timestamp} A timestamp (seconds in current epoch)

  \item{bb} The location of the bounding boxes containing data significantly
  different from the background image.

  \item{mu} The mean (background) image corresponding to this frame

  }
}
\description{
Read frame from a ufmf format movie
}
\details{
When reading multiple frames, the recommended idiom is to read the
  header and store it in an object, then pass that object to read.ufmf along
  with the integer index of the required frame.
}
\examples{
\dontrun{
h=read.ufmf.header("movie.ufmf")
frames=lapply(1:5, function(i) read.ufmf(h, i))
}

}
\seealso{
\code{\link{connection}}, \code{\link{read.ufmf.header}}
}

